Inital Push
this push has the dataset that has been parsed from the aura base code and ready to be used for training as well as the scripts that will help. this was done on an M1 Mac but should be easy enough for any windows or linux users to change and adapt as needed.